Refactor book methods
